Move Resource-related code into its own dir
