+++ src/thing.ts
@@ -46,7 +46,7 @@ import {
   unstable_AclDataset,
 } from "./interfaces";
 import { internal_isAclDataset } from "./acl";
-import { getFetchedFrom } from "./resource";
+import { getFetchedFrom } from "./resource/resource";
 
 /**
  * @hidden Scopes are not yet consistently used in Solid and hence not properly implemented in this library yet (the add*() and set*() functions do not respect it yet), so we're not exposing these to developers at this point in time.
